cleanup: fix unsafe C-style casts and wrap over-long lines
- Replace C-style casts that could silently truncate with explicit
  static_cast, and guard dllexport/API macros behind __INTELLISENSE__
  in vendored headers (exif, imgui, glu-libtess, hidapi, libslic3r).
- Wrap every remaining over-250-char code line (BP1046 part 2):
  function declarations and calls broken at parameter commas, &&/<</+
  chains broken at operators, tbb lambda capture lists wrapped, and
  long string literals split into adjacent literals. Breaks are all at
  separators outside string/comment context, so behavior is unchanged.

Builds clean (0 errors). libslic3r BP1046 is now at 0.

+++ src/libslic3r/AppConfig.cpp
@@ -1241,7 +1241,7 @@ bool AppConfig::set_recent_projects(const std::vector<std::string>& recent_proje
     auto &dst = it_section->second;
 
     std::map<std::string, std::string> src;
-    for (unsigned int i = 0; i < (unsigned int)recent_projects.size(); ++i)
+    for (size_t i = 0; i < recent_projects.size(); ++i)
         src[std::to_string(i + 1)] = recent_projects[i];
 
     if (src != dst) {
